base/rag/indexers/web-docs/requirements.txt
crawl4ai
pymilvus>=2.3
pyyaml>=6.0
httpx>=0.24
